Moved sudden plus calculator stuff to its own package
