+++ app/src/main/AndroidManifest.xml
@@ -9,13 +9,13 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name="com.duraz0rz.iidxtools.activities.SuddenPlusInputActivity">
+        <activity android:name="com.duraz0rz.iidxtools.suddenpluscalculator.activities.SuddenPlusInputActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name="com.duraz0rz.iidxtools.activities.SuddenPlusTableActivity" android:parentActivityName="com.duraz0rz.iidxtools.activities.SuddenPlusInputActivity" />
+        <activity android:name="com.duraz0rz.iidxtools.suddenpluscalculator.activities.SuddenPlusTableActivity" android:parentActivityName="com.duraz0rz.iidxtools.suddenpluscalculator.activities.SuddenPlusInputActivity" />
     </application>
 
 </manifest>
