+++ app/src/main/java/com/duraz0rz/iidxtools/suddenpluscalculator/dataClasses/SuddenPlusDataClasses.kt
@@ -1,3 +1,3 @@
-package com.duraz0rz.iidxtools.dataClasses
+package com.duraz0rz.iidxtools.suddenpluscalculator.dataClasses
 
 data class SuddenPlusValue(val highSpeed : String, val minWhiteNumber: Long, val maxWhiteNumber: Long? = null)
