+++ app/src/main/java/com/duraz0rz/iidxtools/suddenpluscalculator/helpers/SuddenPlusCalculator.kt
@@ -1,6 +1,6 @@
-package com.duraz0rz.iidxtools.helpers
+package com.duraz0rz.iidxtools.suddenpluscalculator.helpers
 
-import com.duraz0rz.iidxtools.dataClasses.SuddenPlusValue
+import com.duraz0rz.iidxtools.suddenpluscalculator.dataClasses.SuddenPlusValue
 
 class SuddenPlusCalculator {
