+++ app/src/main/res/layout/activity_sudden_plus_table.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.duraz0rz.iidxtools.activities.SuddenPlusTableActivity">
+    tools:context="com.duraz0rz.iidxtools.suddenpluscalculator.activities.SuddenPlusTableActivity">
 
     <ScrollView
         android:layout_width="match_parent"
