+++ app/src/test/java/com/duraz0rz/iidxtools/suddenpluscalculator/activities/SuddenPlusInputActivityTest.kt
@@ -1,4 +1,4 @@
-package com.duraz0rz.iidxtools.activities
+package com.duraz0rz.iidxtools.suddenpluscalculator.activities
 
 import android.content.Intent
 import android.widget.Button
